Reports: Implement generating, uploading, downloading and deleting reports
