--- hawk/app/models/report/base.rb
@@ -1,8 +1,0 @@
-# Copyright (c) 2009-2015 Tim Serong <tserong@suse.com>
-# See COPYING for license.
-
-class Report
-  class Base < Tableless
-
-  end
-end
