--- hawk/app/views/explorers/_index.html.erb
@@ -1,205 +1,0 @@
-<% content_for :head do %>
-<style type="text/css">
-/* TODO(should): move to stylesheet */
-#peinputs th {
-  font-weight: bold;
-  padding-left: 1em;
-}
-#peinputs td {
-  padding-left: 1em;
-}
-</style>
-<% end %>
-<div>
-<script type="text/javascript">
-$(function() {
-<% if @hb_report.running? %>
-  setTimeout(function() {
-    // Should this be a GET or POST?
-    $.get("<%= explorer_path %>", {
-      "from_time": "<%= escape_javascript(@from_time) %>",
-      "to_time": "<%= escape_javascript(@to_time) %>",
-      "display": true, "refresh": true }, null, "script");
-  }, 10000);  // Somewhat arbitrary, but this is generally long-running
-<% end %>
-  $("#filter").submit(function() {
-    $(".cache").hide();
-    $("#upload-form").hide();
-    $("#pe-list").children().remove();
-  });
-  $(".cache a.display").click(function(event) {
-    // Clicking any cache link puts chosen times into form, hides all
-    // cached lines and updates the display.  Not quite optimal, but workable.
-    $(".cache").hide();
-    $("#upload-form").hide();
-    $("#from_time").val($(this).parent().prev().prev().prev().text());
-    $("#to_time").val($(this).parent().prev().text());
-  });
-  $(".cache a.delete").click(function(event) {
-    if (!confirm("<%= escape_javascript(_('Delete this history archive?')) %>")) {
-      event.preventDefault();
-      event.stopPropagation();
-      return false;
-    };
-  });
-  $("#filter").submitWithAjax();
-  dialogify_detail_links();  // just in case
-  $("#from_time").datetimepicker({
-    dateFormat: "yy-mm-dd",
-    showOn: "button",
-    buttonImage: url_root + "/assets/icons/calendar.png",
-    buttonImageOnly: true
-  });
-  $("#to_time").datetimepicker({
-    dateFormat: "yy-mm-dd",
-    showOn: "button",
-    buttonImage: url_root + "/assets/icons/calendar.png",
-    buttonImageOnly: true
-  });
-  $("#uploaded_report").change(function(event) {
-    if ($(this).val()) {
-      $("#upload").removeAttr("disabled");
-    } else {
-      $("#upload").attr("disabled", "disabled");
-    }
-  });
-});
-function dialogify_detail_links() {
-  $(".pe-detail").live("click", function(event) {
-    event.preventDefault(); // may not be necessary
-    event.stopPropagation();
-    var peinput = $(this).parent().prev().prev().text();
-    var node = $(this).parent().prev().text();
-    $("#dialog").dialog("option", {
-      width: $(window).width() - 32,
-      height: $(window).height() - 32,
-      position: "center",
-      resizable: true,
-      draggable: true,
-      // TODO(must): localize properly (likewise below)
-      title: peinput + " (" + node + ")"
-    });
-    $("#dialog").html('<div style="text-align: center;" id="dialog-pe-detail">' +
-      '<img src="' + url_root + '/assets/spinner-32x32.gif" alt="" /></div>');
-    $.get($(this).attr("href"), "", function(data) {
-      $("#dialog-pe-detail").css({textAlign: "left"}).html("<pre>" + escape_html(data) + "</pre>");
-    });
-    $("#dialog").dialog("open");
-    return false;
-  });
-  $(".pe-diff").live("click", function(event) {
-    event.preventDefault(); // may not be necessary
-    event.stopPropagation();
-    var peinput = $(this).parent().prev().prev().text();
-    var href = $(this).attr("href") + "&format=html";
-    var args = $.parseQuery(href.substr(href.indexOf("?")));
-    $("#dialog").dialog("option", {
-      width: $(window).width() - 32,
-      height: $(window).height() - 32,
-      position: "center",
-      resizable: true,
-      draggable: true,
-      // TODO(should): localize properly?
-      title: args.left + " : " + args.right
-    });
-    $("#dialog").html('<div style="text-align: center;" id="dialog-pe-diff">' +
-      '<img src="' + url_root + '/assets/spinner-32x32.gif" alt="" /></div>');
-    $.get(href, "", function(data) {
-      $("#dialog-pe-diff").css({textAlign: "left"}).html(data);
-    });
-    $("#dialog").dialog("open");
-    return false;
-  });
-  $(".pe-graph").live("click", function(event) {
-    event.preventDefault();
-    event.stopPropagation();
-    // Note: similar derivation as above, with one more prev :-/
-    var peinput = $(this).parent().prev().prev().prev().text();
-    var node = $(this).parent().prev().prev().text();
-    $("#dialog").dialog("option", {
-      width: $(window).width() - 32,
-      height: $(window).height() - 32,
-      position: "center",
-      resizable: true,
-      draggable: true,
-      // TODO(must): localize properly (likewise above)
-      title: peinput + " (" + node + ")"
-    });
-    if ($(this).attr("href").indexOf("xml") != -1) {
-      // Nasty test, but adequate
-      $("#dialog").html('<div style="text-align: center;" id="dialog-graph-xml">' +
-        '<img src="' + url_root + '/assets/spinner-32x32.gif" alt="" /></div>');
-      $.get($(this).attr("href"), "munge=txt", function(data) {
-        $("#dialog-graph-xml").css({textAlign: "left"}).html("<pre>" + escape_html(data) + "</pre>");
-      });
-    } else {
-      $("#dialog").html('<div style="text-align: center;">' +
-        '<img src="' + $(this).attr("href") + '" alt="" /></div>');
-    }
-    $("#dialog").dialog("open");
-  });
-}
-</script>
-<div style="float: left;">
-<%= form_tag(explorer_path, :method => :get, :id => "filter") do %>
-<%= hidden_field_tag "display", "true" %>
-<table id="filter-form" style="<%= @hb_report.running? ? "display: none;" : "" %>">
-<% unless params[:display] %>
-<% @cache.each do |c| %>
-  <tr class="cache">
-    <td><%=h _("Start Time:") %></td>
-    <td><%=h c[:from_time] %></td>
-    <td><%=h _("End Time:") %></td>
-    <td><%=h c[:to_time] %></td>
-    <td><%= link_to ("Display"), explorer_path(
-      :from_time => c[:from_time], :to_time => c[:to_time],
-      :display => true), :class => "display" %></td>
-    <td><%= link_to ("Delete"), explorer_path(
-      :from_time => c[:from_time], :to_time => c[:to_time],
-      :delete => true), :class => "delete" %></td>
-  </tr>
-<% end %>
-<% end %>
-<% if @upload_name %>
-  <tr>
-    <td><%= _("Exploring %{path}") % { :path => @upload_name } %></td>
-  </tr>
-<% else %>
-  <tr>
-    <th><label for="from_time"><%=h _("Start Time:") %></label></th>
-    <td><%= text_field_tag "from_time", @cache.any? && !params[:display] ? "" : @from_time, :size => 14 %></td>
-    <th><label for="to_time"><%=h _("End Time:") %></label></th>
-    <td><%= text_field_tag "to_time", @cache.any? && !params[:display] ? "" : @to_time, :size => 14 %></td>
-    <td><%= submit_tag(_('Display')) %></td>
-  </tr>
-<% end %>
-</table>
-<% end %>
-<% unless params[:display] %>
-<%= form_tag(explorer_path, :multipart => true) do %>
-<%= hidden_field_tag "display", "true" %>
-<table id="upload-form" style="width: 100%; <%= @hb_report.running? ? "display: none;" : "" %>">
-<tr>
-<td colspan="3"><hr/></td>
-</tr>
-<tr>
-<td><%=h _("Upload hb_report tarball") %></td>
-<td><%= file_field_tag :uploaded_report %></td>
-<td><%= submit_tag _('Explore'), { :id => "upload", :name => "upload", :disabled => true } %></td>
-</tr>
-</table>
-<% end %>
-<% end %>
-<div id="filter-wait" style="<%= @hb_report.running? ? "" : "display: none;" %>">
-<table>
-  <tr style="vertical-align: middle;">
-    <td id="filter-progress"><%=h in_progress_msg %></td>
-    <td><%= image_tag 'spinner-16x16-neutral.gif', :alt => '' %></td>
-  </tr>
-</table>
-</div>
-</div>
-</div>
-<div id="pe-list" style="clear: both;">
-<%= render :partial => "pe_list" %>
-</div>
