--- hawk/app/views/explorers/_index.js.erb
@@ -1,14 +1,0 @@
-<% if @hb_report.running? %>
-$("#filter-form").hide();
-$("#filter-wait").show();
-$("#filter-progress").html(escape_html("<%= escape_javascript(in_progress_msg) %>"));
-setTimeout(function() {
-  // Should this be a GET or POST?
-  $.get("<%= explorer_path %>", {
-    "from_time": "<%= escape_javascript(@from_time) %>",
-    "to_time": "<%= escape_javascript(@to_time) %>",
-    "display": true, "refresh": true }, null, "script");
-}, 10000);  // Somewhat arbitrary, but this is generally long-running
-<% else %>
-window.location.assign("<%=raw explorer_path(:from_time => @from_time, :to_time => @to_time, :display => true) %>");
-<% end %>
