--- hawk/app/views/explorers/_pe_list.html.erb
@@ -1,56 +1,0 @@
-<% if @errors.any? %>
-<div class="ui-state-error ui-corner-all" style="padding: 0.5em; margin-bottom: 0.5em;"><% @errors.each {|e| %><%=h e %><br /><% } %></div>
-<% end %>
-<% if @peinputs %>
-  <% if @peinputs.empty? %>
-<span><%=h _("No transitions found for the specified time period.") %></span>
-  <% else %>
-<div style="float: left; padding-right: 2em;">
-<table id="peinputs" style="white-space: nowrap;">
-  <tr>
-    <th colspan="2"><%=h _('Time') %></th>
-    <th><%=h _('PE Input') %></th>
-    <th><%=h _('Node') %></th>
-  </tr>
-    <%
-      peinputs = @peinputs.reverse
-      for i in 0..peinputs.count-1
-        peinput = peinputs[i]
-    %>
-  <tr>
-    <td><%=h peinput[:timestamp] %></td>
-    <td><span class="ui-icon ui-icon-info " <%=raw(
-      peinput[:info] ? "title=\"#{h peinput[:info]}\"" : 'style="display: none;"'
-    )%>></span></td>
-    <td><%= link_to peinput[:basename], pe_get_path(
-      :from_time => @from_time, :to_time => @to_time, :upload_name => @upload_name,
-      :basename => peinput[:basename], :node => peinput[:node], :path => peinput[:path],
-      :file => "pe-input") %></td>
-    <td><%=h peinput[:node] %></td>
-    <td><%= link_to _("Details"), pe_get_path(
-      :from_time => @from_time, :to_time => @to_time, :upload_name => @upload_name,
-      :basename => peinput[:basename], :node => peinput[:node], :path => peinput[:path],
-      :file => "info"), :target => "hawk-hist-info", :class => "pe-detail" %>
-      <%= link_to _("(full log)"), pe_get_path(
-      :from_time => @from_time, :to_time => @to_time, :upload_name => @upload_name,
-      :basename => peinput[:basename], :node => peinput[:node], :path => peinput[:path],
-      :file => "info", :log => true), :target => "hawk-hist-info", :class => "pe-detail" %></td>
-    <td><%= link_to _("Graph"), pe_get_path(
-      :from_time => @from_time, :to_time => @to_time, :upload_name => @upload_name,
-      :basename => peinput[:basename], :node => peinput[:node], :path => peinput[:path],
-      :file => "graph"), :target => "hawk-hist-info", :class => "pe-graph" %>
-      <%= link_to "(xml)", pe_get_path(
-      :from_time => @from_time, :to_time => @to_time, :upload_name => @upload_name,
-      :basename => peinput[:basename], :node => peinput[:node], :path => peinput[:path],
-      :file => "graph", :format => "xml"), :target => "hawk-hist-info", :class => "pe-graph" %></td>
-    <% if i < peinputs.count-1 %>
-    <td><%= link_to _("Diff"), pe_diff_path(
-      :from_time => @from_time, :to_time => @to_time, :upload_name => @upload_name,
-      :left => peinputs[i+1][:filename], :right => peinput[:filename]), :target => "hawk-hist-info", :class => "pe-diff" %></td>
-    <% end %>
-  </tr>
-    <% end %>
-</table>
-</div>
-  <% end %>
-<% end %>
