--- hawk/app/views/explorers/show.html.haml
@@ -1,55 +1,0 @@
-.container-fluid
-  %h1
-    = icon_text "clock-o", _("History"), class: "page"
-
-  .panel.panel-default
-    .panel-body
-      %ul.nav.nav-pills.nav-justified.iterator
-        %li{ role: "presentation", class: current_page == first_page ? "disabled" : nil }
-          = link_to _("Previous"), url_for(page: prev_page)
-
-        - window_page.each do |page|
-          %li{ role: "presentation", class: current_page == page ? "active" : nil }
-            - if page == "..."
-              .text-center
-                %strong
-                  = page
-            - else
-              = link_to page, url_for(page: page)
-
-        %li{ role: "presentation", class: current_page == last_page ? "disabled" : nil }
-          = link_to _("Next"), url_for(page: next_page)
-
-  .panel.panel-default.with-nav-tabs
-    .panel-heading
-      %ul.nav.nav-tabs.remote
-        %li.active
-          %a{ href: "#details", data: { toggle: "tab", url: detail_explorer_path(id: @report.id, page: current_page) } }
-            = _("Details")
-        %li
-          %a{ href: "#transition", data: { toggle: "tab", url: transition_explorer_path(id: @report.id, page: current_page) } }
-            = _("Transition")
-        %li
-          %a{ href: "#diff", data: { toggle: "tab", url: diff_explorer_path(id: @report.id, page: current_page) } }
-            = _("Diff")
-        %li
-          %a{ href: "#logs", data: { toggle: "tab", url: logs_explorer_path(id: @report.id, page: current_page) } }
-            = _("Logs")
-
-    .panel-body
-      .tab-content
-        #details.tab-pane.fade.in.active
-          .text-center.spacer
-            = icon_tag("spinner", class: "fa-spinner fa-pulse fa-3x")
-
-        #transition.tab-pane.fade
-          .text-center.spacer
-            = icon_tag("spinner", class: "fa-spinner fa-pulse fa-3x")
-
-        #diff.tab-pane.fade
-          .text-center.spacer
-            = icon_tag("spinner", class: "fa-spinner fa-pulse fa-3x")
-
-        #logs.tab-pane.fade
-          .text-center.spacer
-            = icon_tag("spinner", class: "fa-spinner fa-pulse fa-3x")
