--- hawk/app/views/hb_reports/create.js.erb
@@ -1,35 +1,0 @@
-<% if @error %>
-$("#from_time").focus();
-alert("<%= escape_javascript(_("Start Time not specified")) %>");
-<% else %>
-<% if !@hb_report.running? %>
-$("#hbreport-form").show();
-$("#hbreport-wait").hide();
-  <% if @hb_report.lastexit == 0 %>
-    <%
-      # Grab warnings from successful run (e.g.: can't ssh to other node)
-      err = @hb_report.err_filtered
-      if err.any?
-    %>
-    $("#hbreport-status").html("<pre><%= escape_javascript err.join("\n") %></pre>").show();
-    <% else %>
-    $("#hbreport-status").hide();
-    <% end %>
-  $("#hbreport-download").show();
-  $("#hbreport-gentime").text("<%= escape_javascript(File.mtime("/tmp/hb_report-hawk.tar.bz2").strftime("%Y-%m-%d %H:%M:%S")) %>");
-  <% else %>
-  $("#hbreport-status").html("<pre><%= escape_javascript @hb_report.err_lines.join("\n") %></pre>").show();
-  $("#hbreport-download").hide();
-  <% end %>
-<% else %>
-$("#hbreport-form").hide();
-$("#hbreport-status").hide();
-$("#hbreport-download").hide();
-$("#hbreport-wait").show();
-setTimeout(function() {
-  if ($("#dialog").dialog("isOpen")) {
-    $("#dialog").data("req", $.get("<%= hb_reports_status_path %>", "", null, "script"));
-  }
-}, 2000);
-<% end %>
-<% end %>
