--- hawk/app/views/hb_reports/new.html.erb
@@ -1,81 +1,0 @@
-<%
-from_time = params[:from_time]
-if !from_time
-  t = Time.now
-  o = t - 3600
-  if t.day == o.day
-    from_time = o.strftime("%H:%M")
-  else
-    from_time = o.strftime("%Y-%m-%d %H:%M")
-  end
-end
-to_time = params[:to_time]
-if !to_time
-  to_time = Time.now.strftime("%H:%M")
-end
-%>
-<script type="text/javascript">
-$(function() {
-  $($("#generate")[0].form).submitWithAjax();
-  $("#from_time").datetimepicker({
-    dateFormat: "yy-mm-dd",
-    showOn: "button",
-    buttonImage: url_root + "/assets/icons/calendar.png",
-    buttonImageOnly: true
-  });
-  $("#to_time").datetimepicker({
-    dateFormat: "yy-mm-dd",
-    showOn: "button",
-    buttonImage: url_root + "/assets/icons/calendar.png",
-    buttonImageOnly: true
-  });
-  <% if @hb_report.running? %>
-  if ($("#dialog").dialog("isOpen")) {
-    $("#dialog").data("req", $.get("<%= hb_reports_status_path %>", "", null, "script"));
-  }
-  <% else %>
-  $("#hbreport-form").show();
-  <% end %>
-  <% if @hb_report.lastexit == 0 && File.exists?("/tmp/hb_report-hawk.tar.bz2") %>
-    <%
-      # Grab warnings from successful run (e.g.: can't ssh to other node)
-      err = @hb_report.err_filtered
-      if err.any?
-    %>
-    $("#hbreport-status").html("<pre><%= escape_javascript err.join("\n") %></pre>").show();
-    <% else %>
-    $("#hbreport-status").hide();
-    <% end %>
-  $("#hbreport-download").show();
-  $("#hbreport-gentime").text("<%= escape_javascript(File.mtime("/tmp/hb_report-hawk.tar.bz2").strftime("%Y-%m-%d %H:%M:%S")) %>");
-  <% end %>
-});
-</script>
-<div id="hbreport-form" style="display: none;">
-<%= form_tag(:action => "create") do %>
-<table style="width: 100%;">
-  <tr>
-    <th><label for="from_time"><%=h _("Start Time:") %></label></th>
-    <td><%= text_field_tag "from_time", from_time, :size => 14 %></td>
-    <td style="text-align: right;"><%= submit_tag(_("Generate"), { :id => "generate" }) %></td>
-  </tr>
-  <tr>
-    <th><label for="to_time"><%=h _("End Time:") %></label></th>
-    <td><%= text_field_tag "to_time", to_time, :size => 14 %></td>
-  </tr>
-</table>
-<% end %>
-</div>
-<div id="hbreport-status" style="display: none; font-size: 80%;">
-</div>
-<div id="hbreport-wait" style="display: none;">
-<table style="width: 100%;">
-  <tr style="vertical-align: middle;">
-    <td><%=h _("hb_report generation in progress...") %></td>
-    <td><%= image_tag 'spinner-16x16-neutral.gif', :alt => '' %></td>
-  </tr>
-</table>
-</div>
-<div id="hbreport-download" style="display: none; padding-top: 1em;">
-<%= link_to _("Download hb_report"), hb_report_path(:id => "hb_report-hawk") %> (<span id="hbreport-gentime"></span>)
-</div>
