--- hawk/app/views/hb_reports/permission_denied.html.erb
@@ -1,3 +1,0 @@
-<div class="ui-state-error ui-corner-all" style="padding: 0.5em; margin-bottom: 0.5em;">
-<%=h _("Error: Permission denied for user '%s'.") % current_user %>
-</div>
